// components/(dashboards)/admin-dashboard/analytics/enums.ts
// Enums for Analytics

export enum TabType {
  OVERVIEW = "Overview",
  USERS = "Users",
  COOPERATIVES = "Cooperatives",
  REVENUE = "Revenue"
}

export enum TimeFilter {
  THIS_MONTH = "This Month",
  LAST_MONTH = "Last Month",
  THIS_QUARTER = "This Quarter",
  THIS_YEAR = "This Year"
}